crawler/requirements.txt
kafka-python
pymongo
scrapy
bs4
requests
pattern